udpate code

+++ mcquic/validate/cli.py
@@ -3,12 +3,12 @@ import click
 import pathlib
 import logging
 
+import mcquic
+
 
 def checkArgs(debug: bool, quiet: bool, path: pathlib.Path, output: pathlib.Path):
     if path.is_dir():
         raise ValueError("Please provide a file path to `path`, not a dir.")
-    if output.is_dir() and not output.exists():
-        raise ValueError("`output` dir does not exist.")
     if quiet:
         return logging.CRITICAL
     if debug:
@@ -18,9 +18,6 @@ def checkArgs(debug: bool, quiet: bool, path: pathlib.Path, output: pathlib.Path
 
 def main(debug: bool, quiet: bool, path: pathlib.Path, images: pathlib.Path, output: pathlib.Path) -> int:
     loggingLevel = checkArgs(debug, quiet, path, output)
-
-    import shutil, gzip
-
     import torch
     from vlutils.logger import configLogging
 
@@ -39,7 +36,15 @@ def main(debug: bool, quiet: bool, path: pathlib.Path, images: pathlib.Path, out
 
     model = Compressor(**config.Model.Params).cuda()
 
-    modelStateDict = {key[len("module._compressor."):]: value for key, value in checkpoint["trainer"]["_model"].items()}
+    if "trainer" in checkpoint:
+        modelStateDict = {key[len("module._compressor."):]: value for key, value in checkpoint["trainer"]["_model"].items()}
+    else:
+        modelStateDict = checkpoint["model"]
+        logger.warning("I got an already-converted ckpt. The `output` will be ignored. If you still want to export this ckpt, please copy it directly.")
+        if not "version" in checkpoint or (checkpoint["version"] != mcquic.__version__):
+            v = checkpoint.get("version", None)
+            logger.warning(f"Version mismatch: It seems this ckpt has a version {v} but mcquic now is {mcquic.__version__}.")
+        output = None
 
     model.load_state_dict(modelStateDict) # type: ignore
 
@@ -50,19 +55,24 @@ def main(debug: bool, quiet: bool, path: pathlib.Path, images: pathlib.Path, out
     progress = getRichProgress()
 
     with progress:
-        _, summary = validator.validate(checkpoint["trainer"]["_epoch"], model, valLoader, progress)
+        _, summary = validator.validate(None, model, valLoader, progress)
         logger.info(summary)
-        _, speedSummary = validator.speed(checkpoint["trainer"]["_epoch"], model, progress)
+        _, speedSummary = validator.speed(None, model, progress)
         logger.info(speedSummary)
+
+    if output is None or (output.is_dir() and not output.exists()):
+        logger.info(f"I got an invalid path: `{output}`, skip saving model.")
+        return 0
 
     if output.is_dir():
         modelName = "_".join([f"{key}_{value}" for key, value in config.Model.params.items()])
         modelName = modelName.replace(", ", "_").replace("[", "").replace("]", "")
-        output = output.joinpath(f"{modelName}_{config.Training.Target.lower()}.mcquic")
+        output = output.joinpath(f"{modelName}_{config.Train.Target.lower()}.mcquic")
 
     torch.save({
         "model": model.state_dict(),
-        "config": config.serialize()
+        "config": config.serialize(),
+        "version": mcquic.__version__
     }, output)
 
     logger.info(f"Saved at `{output}`.")
@@ -75,7 +85,7 @@ def main(debug: bool, quiet: bool, path: pathlib.Path, images: pathlib.Path, out
 @click.option("-q", "--quiet", is_flag=True, help="Silence all messages, this option has higher priority to `-D/--debug`.")
 @click.argument("path", type=click.Path(exists=True, dir_okay=False, resolve_path=True, path_type=pathlib.Path), required=True, nargs=1)
 @click.argument("images", type=click.Path(exists=True, file_okay=False, resolve_path=True, path_type=pathlib.Path), required=True, nargs=1)
-@click.argument("output", type=click.Path(exists=False, dir_okay=True, resolve_path=True, path_type=pathlib.Path), required=True, nargs=1)
+@click.argument("output", type=click.Path(exists=False, dir_okay=True, resolve_path=True, path_type=pathlib.Path), required=False, nargs=1)
 def entryPoint(debug, quiet, path, images, output):
     """Validate a trained model from `path` by images from `images` dir, and publish a final state_dict to `output` path.
